windows: Consume reader immediately when moving out handle
When moving the associated handle to another process, immediately
consume the `MessageReader` containing it, so we do not leave the reader
hanging around with an invalid handle.

While this doesn't really change much in practice -- the reader was
dropped along with the receiver shortly after anyway -- it's cleaner
semantically, and should be a tick more robust.