windows: Don't explicitly set `ErrorKind::BrokenPipe` on closed sender
This is related to the fix for the `inprocess` back-end in
750d9a12512727e31c9e9bc0616d271a106d7a57 : according to the
documentation of `ErrorKind`, just like the related Unix error,
`ErrorKind::BrokenPipe` is supposed to signal only a "receiver closed"
condition -- and not "sender closed".

Note that the situation is quite confusing here, since the Windows API
actually returns a `winapi::ERROR_BROKEN_PIPE` for the "sender closed"
condition; so it would seem an obvious choice to turn it into
`ErrorKind::BrokenPipe`... Except that this conflicts with the stated
purpose according to the documentation.

Rather than explicitly deciding on the right `ErrorKind` to use here, we
just punt this back to the Windows API, letting it assign whatever
`ErrorKind` it deems appropriate for this situation -- thus avoiding a
potential confusing discrepancy with other `winapi` users, at the cost
of a potential confusing discrepancy with other `ipc-channel`
back-ends...

The situation here differs from that in the `inprocess` back-end also in
that we were already using a distinct `WinError::ChannelClosed` internal
state for closed senders, rather than throwing it in with the handling
of closed receivers -- so it didn't affect the behaviour of the public
`channel_is_closed()` method here, but only the converted `Error` value.
(And thus the message printed on `unwrap()`.)