windows: Don't delay "sender closed" notification from `select()`
Immediately handle the `closed` status -- rather than postponing
handling to the next `select()` call -- even if it gets set in
`start_read()`.

(It already was being reported immediately in case the status got set in
`notify_completion()`.)

This is undoubtedly the more obvious and desirable behaviour. (I wonder
whether we should actually enforce it with a test case?...) Also, the
code should be easier to follow this way.